Pull in text to summarize principles
This is what I’ve written based on our discussion the week before IETF 110, as an attempt to write down the high level guidance and principles we could provide (from my perspective). I’m not sure the best way that this meshes with the current structure, but wanted to share to start a discussion.