New abstract
This partly address #21.

The second paragraph still talks a lot about encryption. I personally think this is a good thing, but we can also cut at least the last sentence if others think that's better...?

There is also issue #23 which is about layers in the asbtract. The abstract still talks about layers but I'm fine to remove that if we agree we want to; not sure we agreed on this issue yet...?