Merge intro part of principle section into actual intro
fixes #29

I kept most of the text in the principles section and moved to the respective place in the intro. That makes the intro a bit long now, however, I think this fits nicely together and rather keep all the text... or we can still take another pass to cut if needed later.